requirements.txt
flask
openai>=1.0.0
python-dotenv
tiktoken
werkzeug
PyPDF2
firebase-admin
pyrebase4
gunicorn